requirements.txt
numpy>=1.26.0
pandas>=2.2.0
pyyaml>=6.0.1
tqdm>=4.66.0
datasets>=2.18.0
evaluate>=0.4.1
rank-bm25>=0.2.2
matplotlib>=3.8.0
seaborn>=0.13.0
openai>=1.14.0
python-dotenv>=1.0.1
pytest>=8.0.0
